feat: Process Deep Link
Treat the new "notificationClick" event, triggered when the user clicks a Push Notification that contains a deep link. In order to work properly when the app gets launched from clicking a Push Notification, add two new control properties and one method, that checks if the device is ready to trigger events and that postpones it into an event queue when it's not.